backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
databases==0.9.0
distro==1.9.0
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.2.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.60
langchain-openai==0.3.17
langsmith==0.3.42
librosa==0.10.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
onnxruntime==1.21.0
openai==1.79.0
openai-whisper==20240930
orjson==3.10.18
packaging==24.2
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
propcache==0.3.1
protobuf==6.30.2
pyannote.audio==3.1.1
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.1
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.9.0
torch==2.6.0
transformers==4.37.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.19.0
zstandard==0.23.0
langchain-openai>=0.0.1
ace_tools==0.0
asttokens==3.0.0
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
lightgbm==4.6.0
matplotlib-inline==0.1.7
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pytz==2025.2
pywin32==310
pyzmq==26.4.0
scikit-learn==1.6.1
scipy==1.15.3
stack-data==0.6.3
threadpoolctl==3.6.0
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
